Fixed PHP-773: Fix the persistence issue with logging stream contexts
Each time we instanciate MongoClient we will iterate over open servers
and apply the current stream contexts.
This will overwrite previously set contexts with the new one or reset it
to NULL if no context was provided.

If there are no open streams, we assign the context to the server
options and it will be used when creating new streams like before